refactor: split addExerciseDialog into custom hook
