+++ SparkyFitnessFrontend/src/pages/Exercises/AddExerciseDialog.tsx
@@ -8,30 +8,17 @@ import {
   DialogDescription,
 } from '@/components/ui/dialog';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
+
 import ExerciseSearch from './ExerciseSearch';
 import WorkoutPresetSelector from './WorkoutPresetSelector';
 import ExerciseImportCSV, { type ExerciseCSVData } from './ExerciseImportCSV';
 import ExerciseEntryHistoryImportCSV from './ExerciseEntryHistoryImportCSV';
 import type { WorkoutPreset } from '@/types/workout';
-import { useAuth } from '@/hooks/useAuth';
 import { toast } from '@/hooks/use-toast';
-import { XCircle } from 'lucide-react';
-import {
-  useCreateExerciseMutation,
-  useImportExercisesJsonMutation,
-} from '@/hooks/Exercises/useExercises';
+import { useImportExercisesJsonMutation } from '@/hooks/Exercises/useExercises';
 import { Exercise } from '@/types/exercises';
+import { useAddCustomExerciseForm } from '@/hooks/Exercises/useAddCustomExerciseForm';
+import AddCustomExerciseForm from './AddCustomExerciseForm';
 
 interface ImportConflictError {
   status?: number;
@@ -59,36 +46,10 @@ const AddExerciseDialog = ({
   onWorkoutPresetSelected,
 }: AddExerciseDialogProps) => {
   const { t } = useTranslation();
-  const { user } = useAuth();
+  const customForm = useAddCustomExerciseForm(onExerciseAdded, onOpenChange);
   const [activeTab, setActiveTab] = useState(
     mode === 'database-manager' ? 'online' : 'my-exercises'
   );
-  const [newExerciseName, setNewExerciseName] = useState('');
-  const [newExerciseCategory, setNewExerciseCategory] = useState('general');
-  const [newExerciseCalories, setNewExerciseCalories] = useState(300); // Default calculated calories
-  const [manualCaloriesPerHour, setManualCaloriesPerHour] = useState<
-    number | undefined
-  >(undefined); // For manual override
-  const [newExerciseDescription, setNewExerciseDescription] = useState('');
-  const [newExerciseSource, setNewExerciseSource] = useState('custom'); // Default to "custom"
-  const [newExerciseForce, setNewExerciseForce] = useState('');
-  const [newExerciseLevel, setNewExerciseLevel] = useState('');
-  const [newExerciseMechanic, setNewExerciseMechanic] = useState('');
-  const [newExerciseEquipment, setNewExerciseEquipment] = useState('');
-  const [newExercisePrimaryMuscles, setNewExercisePrimaryMuscles] =
-    useState('');
-  const [newExerciseSecondaryMuscles, setNewExerciseSecondaryMuscles] =
-    useState('');
-  const [newExerciseInstructions, setNewExerciseInstructions] = useState('');
-  const [newExerciseImages, setNewExerciseImages] = useState<File[]>([]); // State to hold image files
-  const [newExerciseImageUrls, setNewExerciseImageUrls] = useState<string[]>(
-    []
-  ); // State to hold image URLs for display
-  const [draggedImageIndex, setDraggedImageIndex] = useState<number | null>(
-    null
-  ); // For reordering
-
-  const { mutateAsync: createExercise } = useCreateExerciseMutation();
   const { mutateAsync: importExerciseFromJson } =
     useImportExercisesJsonMutation();
 
@@ -105,149 +66,6 @@ const AddExerciseDialog = ({
     });
     onExerciseAdded(exercise, sourceMode); // Pass the selected exercise and source mode
     onOpenChange(false);
-  };
-
-  const handleAddCustomExercise = async () => {
-    if (!user) return;
-    try {
-      const newExercise = {
-        name: newExerciseName,
-        category: newExerciseCategory,
-        calories_per_hour:
-          manualCaloriesPerHour !== undefined
-            ? manualCaloriesPerHour
-            : newExerciseCalories, // Use manual if provided
-        description: newExerciseDescription,
-        user_id: user.id,
-        is_custom: true,
-        source: newExerciseSource,
-        force: newExerciseForce,
-        level: newExerciseLevel,
-        mechanic: newExerciseMechanic,
-        equipment: newExerciseEquipment
-          .split(',')
-          .map((s) => s.trim())
-          .filter((s) => s),
-        primary_muscles: newExercisePrimaryMuscles
-          .split(',')
-          .map((s) => s.trim())
-          .filter((s) => s),
-        secondary_muscles: newExerciseSecondaryMuscles
-          .split(',')
-          .map((s) => s.trim())
-          .filter((s) => s),
-        instructions: newExerciseInstructions
-          .split('\n')
-          .map((s) => s.trim())
-          .filter((s) => s),
-        // images: newExerciseImageUrls, // Do not send URLs to backend, server will handle based on uploaded files
-      };
-
-      const formData = new FormData();
-      formData.append('exerciseData', JSON.stringify(newExercise));
-      newExerciseImages.forEach((file) => {
-        formData.append('images', file);
-      });
-
-      const createdExercise = await createExercise(formData);
-      toast({
-        title: t('common.success', 'Success'),
-        description: t(
-          'exercise.addExerciseDialog.addSuccess',
-          'Exercise added successfully'
-        ),
-      });
-
-      onExerciseAdded(createdExercise, 'custom');
-      onOpenChange(false);
-      setNewExerciseName('');
-      setNewExerciseCategory('general');
-      setNewExerciseCalories(300);
-      setNewExerciseDescription('');
-      setNewExerciseSource('custom');
-      setNewExerciseForce('');
-      setNewExerciseLevel('');
-      setNewExerciseMechanic('');
-      setNewExerciseEquipment('');
-      setNewExercisePrimaryMuscles('');
-      setNewExerciseSecondaryMuscles('');
-      setNewExerciseInstructions('');
-      setNewExerciseImages([]);
-      setNewExerciseImageUrls([]);
-    } catch (error) {
-      console.error('Error adding exercise:', error);
-      toast({
-        title: t('common.error', 'Error'),
-        description: t(
-          'exercise.addExerciseDialog.addError',
-          'Failed to add exercise'
-        ),
-        variant: 'destructive',
-      });
-    }
-  };
-
-  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files) {
-      const filesArray = Array.from(e.target.files);
-      setNewExerciseImages((prevImages) => [...prevImages, ...filesArray]);
-      filesArray.forEach((file) => {
-        const reader = new FileReader();
-        reader.onloadend = () => {
-          setNewExerciseImageUrls((prevUrls) => [
-            ...prevUrls,
-            reader.result as string,
-          ]);
-        };
-        reader.readAsDataURL(file);
-      });
-    }
-  };
-
-  const handleRemoveImage = (indexToRemove: number) => {
-    setNewExerciseImages((prevImages) =>
-      prevImages.filter((_, index) => index !== indexToRemove)
-    );
-    setNewExerciseImageUrls((prevUrls) =>
-      prevUrls.filter((_, index) => index !== indexToRemove)
-    );
-  };
-
-  const handleDragStart = (
-    e: React.DragEvent<HTMLDivElement>,
-    index: number
-  ) => {
-    setDraggedImageIndex(index);
-    e.dataTransfer.effectAllowed = 'move';
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.dataTransfer.dropEffect = 'move';
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>, index: number) => {
-    e.preventDefault();
-    if (draggedImageIndex === null || draggedImageIndex === index) {
-      return;
-    }
-
-    const reorderedImages = [...newExerciseImages];
-    const reorderedUrls = [...newExerciseImageUrls];
-
-    const [draggedFile] = reorderedImages.splice(draggedImageIndex, 1);
-    const [draggedUrl] = reorderedUrls.splice(draggedImageIndex, 1);
-
-    if (draggedFile) {
-      reorderedImages.splice(index, 0, draggedFile);
-    }
-    if (draggedUrl) {
-      reorderedUrls.splice(index, 0, draggedUrl);
-    }
-
-    setNewExerciseImages(reorderedImages);
-    setNewExerciseImageUrls(reorderedUrls);
-    setDraggedImageIndex(null);
   };
 
   const handleImportFromCSV = async (
@@ -287,7 +105,7 @@ const AddExerciseDialog = ({
         className={
           activeTab === 'import-csv' || activeTab === 'import-history-csv'
             ? 'sm:max-w-[95vw] sm:max-h-[95vh] w-[95vw] h-[95vh] overflow-y-auto'
-            : 'sm:max-w-[800px] overflow-y-auto max-h-[90vh]'
+            : 'sm:max-w-200 overflow-y-auto max-h-[90vh]'
         }
       >
         <DialogHeader>
@@ -359,305 +177,7 @@ const AddExerciseDialog = ({
             </div>
           </TabsContent>
           <TabsContent value="custom" className="overflow-y-auto max-h-full">
-            <div className="grid gap-4 py-4">
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="name" className="text-right">
-                  {t('exercise.addExerciseDialog.nameLabel', 'Name')}
-                </Label>
-                <Input
-                  id="name"
-                  value={newExerciseName}
-                  onChange={(e) => setNewExerciseName(e.target.value)}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="category" className="text-right">
-                  {t('exercise.addExerciseDialog.categoryLabel', 'Category')}
-                </Label>
-                <Select
-                  onValueChange={setNewExerciseCategory}
-                  defaultValue={newExerciseCategory}
-                >
-                  <SelectTrigger className="col-span-3">
-                    <SelectValue
-                      placeholder={t(
-                        'exercise.addExerciseDialog.selectCategoryPlaceholder',
-                        'Select a category'
-                      )}
-                    />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="general">
-                      {t(
-                        'exercise.addExerciseDialog.categoryGeneral',
-                        'General'
-                      )}
-                    </SelectItem>
-                    <SelectItem value="strength">
-                      {t(
-                        'exercise.addExerciseDialog.categoryStrength',
-                        'Strength'
-                      )}
-                    </SelectItem>
-                    <SelectItem value="cardio">
-                      {t('exercise.addExerciseDialog.categoryCardio', 'Cardio')}
-                    </SelectItem>
-                    <SelectItem value="yoga">
-                      {t('exercise.addExerciseDialog.categoryYoga', 'Yoga')}
-                    </SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="calories" className="text-right">
-                  {t(
-                    'exercise.addExerciseDialog.caloriesPerHourLabel',
-                    'Calories/Hour'
-                  )}
-                </Label>
-                <Input
-                  id="calories"
-                  type="number"
-                  value={
-                    manualCaloriesPerHour !== undefined
-                      ? manualCaloriesPerHour.toString()
-                      : newExerciseCalories.toString()
-                  }
-                  onChange={(e) =>
-                    setManualCaloriesPerHour(Number(e.target.value))
-                  }
-                  placeholder="Calculated: 300" // Show calculated as placeholder
-                  className="col-span-3"
-                />
-                <p className="col-span-4 text-xs text-muted-foreground">
-                  {t(
-                    'exercise.addExerciseDialog.caloriesPerHourHint',
-                    'Leave blank to use system calculated calories per hour ({{newExerciseCalories}} cal/hour).',
-                    { newExerciseCalories }
-                  )}
-                </p>
-              </div>
-              <div className="grid grid-cols-4 items-start gap-4">
-                <Label htmlFor="description" className="text-right mt-1">
-                  {t(
-                    'exercise.addExerciseDialog.descriptionLabel',
-                    'Description'
-                  )}
-                </Label>
-                <Textarea
-                  id="description"
-                  value={newExerciseDescription}
-                  onChange={(e) => setNewExerciseDescription(e.target.value)}
-                  className="col-span-3"
-                />
-              </div>
-            </div>
-            <div className="grid grid-cols-4 items-center gap-4">
-              <Label htmlFor="source" className="text-right">
-                {t('exercise.addExerciseDialog.sourceLabel', 'Source')}
-              </Label>
-              <Input
-                id="source"
-                value={newExerciseSource}
-                onChange={(e) => setNewExerciseSource(e.target.value)}
-                className="col-span-3"
-              />
-            </div>
-            <div className="grid grid-cols-4 items-center gap-4">
-              <Label htmlFor="force" className="text-right">
-                {t('exercise.addExerciseDialog.forceLabel', 'Force')}
-              </Label>
-              <Select
-                onValueChange={setNewExerciseForce}
-                defaultValue={newExerciseForce}
-              >
-                <SelectTrigger className="col-span-3">
-                  <SelectValue
-                    placeholder={t(
-                      'exercise.addExerciseDialog.selectForcePlaceholder',
-                      'Select force'
-                    )}
-                  />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="pull">
-                    {t('exercise.addExerciseDialog.forcePull', 'Pull')}
-                  </SelectItem>
-                  <SelectItem value="push">
-                    {t('exercise.addExerciseDialog.forcePush', 'Push')}
-                  </SelectItem>
-                  <SelectItem value="static">
-                    {t('exercise.addExerciseDialog.forceStatic', 'Static')}
-                  </SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="grid grid-cols-4 items-center gap-4">
-              <Label htmlFor="level" className="text-right">
-                {t('exercise.addExerciseDialog.levelLabel', 'Level')}
-              </Label>
-              <Select
-                onValueChange={setNewExerciseLevel}
-                defaultValue={newExerciseLevel}
-              >
-                <SelectTrigger className="col-span-3">
-                  <SelectValue
-                    placeholder={t(
-                      'exercise.addExerciseDialog.selectLevelPlaceholder',
-                      'Select level'
-                    )}
-                  />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="beginner">
-                    {t('exercise.addExerciseDialog.levelBeginner', 'Beginner')}
-                  </SelectItem>
-                  <SelectItem value="intermediate">
-                    {t(
-                      'exercise.addExerciseDialog.levelIntermediate',
-                      'Intermediate'
-                    )}
-                  </SelectItem>
-                  <SelectItem value="expert">
-                    {t('exercise.addExerciseDialog.levelExpert', 'Expert')}
-                  </SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="grid grid-cols-4 items-center gap-4">
-              <Label htmlFor="mechanic" className="text-right">
-                {t('exercise.addExerciseDialog.mechanicLabel', 'Mechanic')}
-              </Label>
-              <Select
-                onValueChange={setNewExerciseMechanic}
-                defaultValue={newExerciseMechanic}
-              >
-                <SelectTrigger className="col-span-3">
-                  <SelectValue
-                    placeholder={t(
-                      'exercise.addExerciseDialog.selectMechanicPlaceholder',
-                      'Select mechanic'
-                    )}
-                  />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="isolation">
-                    {t(
-                      'exercise.addExerciseDialog.mechanicIsolation',
-                      'Isolation'
-                    )}
-                  </SelectItem>
-                  <SelectItem value="compound">
-                    {t(
-                      'exercise.addExerciseDialog.mechanicCompound',
-                      'Compound'
-                    )}
-                  </SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="grid grid-cols-4 items-start gap-4">
-              <Label htmlFor="equipment" className="text-right mt-1">
-                {t(
-                  'exercise.addExerciseDialog.equipmentLabel',
-                  'Equipment (comma-separated)'
-                )}
-              </Label>
-              <Textarea
-                id="equipment"
-                value={newExerciseEquipment}
-                onChange={(e) => setNewExerciseEquipment(e.target.value)}
-                className="col-span-3"
-              />
-            </div>
-            <div className="grid grid-cols-4 items-start gap-4">
-              <Label htmlFor="primaryMuscles" className="text-right mt-1">
-                {t(
-                  'exercise.addExerciseDialog.primaryMusclesLabel',
-                  'Primary Muscles (comma-separated)'
-                )}
-              </Label>
-              <Textarea
-                id="primaryMuscles"
-                value={newExercisePrimaryMuscles}
-                onChange={(e) => setNewExercisePrimaryMuscles(e.target.value)}
-                className="col-span-3"
-              />
-            </div>
-            <div className="grid grid-cols-4 items-start gap-4">
-              <Label htmlFor="secondaryMuscles" className="text-right mt-1">
-                {t(
-                  'exercise.addExerciseDialog.secondaryMusclesLabel',
-                  'Secondary Muscles (comma-separated)'
-                )}
-              </Label>
-              <Textarea
-                id="secondaryMuscles"
-                value={newExerciseSecondaryMuscles}
-                onChange={(e) => setNewExerciseSecondaryMuscles(e.target.value)}
-                className="col-span-3"
-              />
-            </div>
-            <div className="grid grid-cols-4 items-start gap-4">
-              <Label htmlFor="instructions" className="text-right mt-1">
-                {t(
-                  'exercise.addExerciseDialog.instructionsLabel',
-                  'Instructions (one per line)'
-                )}
-              </Label>
-              <Textarea
-                id="instructions"
-                value={newExerciseInstructions}
-                onChange={(e) => setNewExerciseInstructions(e.target.value)}
-                className="col-span-3"
-              />
-            </div>
-            <div className="grid grid-cols-4 items-start gap-4">
-              <Label htmlFor="images" className="text-right mt-1">
-                {t('exercise.addExerciseDialog.imagesLabel', 'Images')}
-              </Label>
-              <div className="col-span-3">
-                <Input
-                  id="images"
-                  type="file"
-                  multiple
-                  accept="image/*"
-                  onChange={handleImageChange}
-                  className="col-span-3"
-                />
-                <div className="mt-2 flex flex-wrap gap-2">
-                  {newExerciseImageUrls.map((url, index) => (
-                    <div
-                      key={index}
-                      draggable
-                      onDragStart={(e) => handleDragStart(e, index)}
-                      onDragOver={handleDragOver}
-                      onDrop={(e) => handleDrop(e, index)}
-                      className="relative w-24 h-24 cursor-grab"
-                    >
-                      <img
-                        src={url}
-                        alt={`preview ${index}`}
-                        className="w-full h-full object-cover rounded"
-                      />
-                      <Button
-                        type="button"
-                        variant="destructive"
-                        size="icon"
-                        className="absolute -top-2 -right-2 h-6 w-6 rounded-full"
-                        onClick={() => handleRemoveImage(index)}
-                      >
-                        <XCircle className="h-4 w-4" />
-                      </Button>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
-            <Button onClick={handleAddCustomExercise}>
-              {t('exercise.addExerciseDialog.addButton', 'Add Exercise')}
-            </Button>
+            <AddCustomExerciseForm form={customForm} />
           </TabsContent>
           <TabsContent value="import-csv">
             <div className="pt-4">
